Various bug fixes

+++ code/mas_princess/src/main/java/env/objects/structures/Bridge.java
@@ -3,14 +3,14 @@ import env.utils.*;
 
 // Bridge structure
 public class Bridge extends MapStructure {
-    private final double slipProbability;
+    private final int slipProbability;
 
-    public Bridge(double slipProbability, Pose pose) {
+    public Bridge(int slipProbability, Pose pose) {
         super("bridge", false , false, true, 1, null, pose); // Bridges are breakable, not repairable, 3 cells wide, 2 cells tall, and have 1 life point
         this.slipProbability = slipProbability;
     }
 
-    public double getSlipProbability () {
+    public int getSlipProbability () {
         return this.slipProbability;
     }
